2. back-end/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
gunicorn==21.2.0
python-dotenv==1.0.0
pydantic==2.4.2
requests==2.31.0
python-jose==3.3.0
supabase==1.0.3
httpx==0.24.1